% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_templates.R
\name{era_copc_temp}
\alias{era_copc_temp}
\title{Templates for creating tables}
\usage{
era_copc_temp(file_path)
}
\arguments{
\item{file_path}{text string of file path to save template}
}
\description{
Templates for creating tables
}
