% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_row_height.R
\name{excel_row_height}
\alias{excel_row_height}
\title{Fixes row height if sub/superscripts are present in table}
\usage{
excel_row_height(strings)
}
\arguments{
\item{strings}{sharedStrings from an openxlsx workbook object}
}
\value{
strings for an openxlsx workbook object
}
\description{
Fixes row height if sub/superscripts are present in table
}
\examples{
wb <- openxlsx::createWorkbook()

openxlsx::addWorksheet(wb, "sheet")

excel_row_height(strings = wb$sharedStrings)
}
