% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_superscript.R
\name{excel_superscript}
\alias{excel_superscript}
\title{Adds superscripts to text in excel}
\usage{
excel_superscript(strings)
}
\arguments{
\item{strings}{sharedStrings from an openxlsx workbook object}
}
\value{
strings for an openxlsx workbook object
}
\description{
Adds superscripts to text in excel
}
\examples{
wb <- openxlsx::createWorkbook()

openxlsx::addWorksheet(wb, "sheet")

excel_superscript(strings = wb$sharedStrings)
}
