% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_abbr.R
\name{fmt_abbr}
\alias{fmt_abbr}
\title{Adds abbreviations to an excel file}
\usage{
fmt_abbr(wb, sheet, abbr_text, start_row = 1, start_col = 1)
}
\arguments{
\item{wb}{an openxlsx workbook object}

\item{sheet}{name of sheet in workbook object}

\item{abbr_text}{dataframe of text to display in abbreviations section of
workbook object. Include "MMMMM_#" to merge # number of cells to the right.}

\item{start_row}{number for starting row}

\item{start_col}{number for starting column}
}
\value{
A list with the updated workbook object (wb) and number of the last
row in the updated workbook object (length).
}
\description{
Adds abbreviations to an excel file
}
\examples{
wb <- openxlsx::createWorkbook()

openxlsx::addWorksheet(wb, "sheet")

fmt_abbr(wb = wb, sheet = "sheet",
abbr_text = abbr_text,
start_row = 2, start_col = 1)
}
