% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_columns.R
\name{fmt_columns}
\alias{fmt_columns}
\title{Adds column headers to an excel file}
\usage{
fmt_columns(wb, sheet, col_names, start_row = 1, start_col = 1, end_col)
}
\arguments{
\item{wb}{an openxlsx workbook object}

\item{sheet}{name of sheet in workbook object}

\item{col_names}{dataframe of text to display in column headers of
workbook object}

\item{start_row}{number for starting row}

\item{start_col}{number for starting column}

\item{end_col}{number for the last column}
}
\value{
A list with the updated workbook object (wb) and number of the last
row in the updated workbook object (length)
}
\description{
Adds column headers to an excel file
}
\examples{
wb <- openxlsx::createWorkbook()

openxlsx::addWorksheet(wb, "sheet")

fmt_columns(wb = wb, sheet = "sheet",
col_names = col_names,
start_row = 2, start_col = 1, end_col = 17)
}
