% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_footnote.R
\name{fmt_footnote}
\alias{fmt_footnote}
\title{Adds footnotes to an excel file}
\usage{
fmt_footnote(wb, sheet, footnote_text, start_row = 1, start_col = 1, end_col)
}
\arguments{
\item{wb}{an openxlsx workbook object}

\item{sheet}{name of sheet in workbook object}

\item{footnote_text}{dataframe of text to display in footnote section of
workbook object}

\item{start_row}{number for starting row}

\item{start_col}{number for starting column}

\item{end_col}{number for the last column}
}
\value{
A list with the updated workbook object (wb) and number of the last
row in the updated workbook object (length).
}
\description{
Adds footnotes to an excel file
}
\examples{
wb <- openxlsx::createWorkbook()

openxlsx::addWorksheet(wb, "sheet")

fmt_footnote(wb = wb, sheet = "sheet",
footnote_text = foot_text,
start_row = 2, start_col = 1, end_col = 17)
}
