% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_header.R
\name{fmt_header}
\alias{fmt_header}
\title{Adds a header to an excel file}
\usage{
fmt_header(wb, sheet, header_text, start_row = 1, start_col = 1, end_col)
}
\arguments{
\item{wb}{an openxlsx workbook object}

\item{sheet}{name of sheet in workbook object}

\item{header_text}{text string to display in header of workbook object}

\item{start_row}{number for starting row}

\item{start_col}{number for starting column}

\item{end_col}{number for the last column}
}
\value{
A list with the updated workbook object (wb) and number of the last
row in the updated workbook object (length)
}
\description{
Adds a header to an excel file
}
\examples{
wb <- openxlsx::createWorkbook()

openxlsx::addWorksheet(wb, "sheet")

fmt_header(wb = wb, sheet = "sheet",
header_text = "DRAFT - Privileged and Confidential",
start_row = 1, start_col = 1, end_col = 17)
}
