% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_ref.R
\name{fmt_ref}
\alias{fmt_ref}
\title{Adds references to an excel file}
\usage{
fmt_ref(wb, sheet, ref_text, start_row = 1, start_col = 1, end_col)
}
\arguments{
\item{wb}{an openxlsx workbook object}

\item{sheet}{name of sheet in workbook object}

\item{ref_text}{dataframe of text (column name 'reference_text') to display
in reference section of workbook object}

\item{start_row}{number for starting row}

\item{start_col}{number for starting column}

\item{end_col}{number for the last column}
}
\value{
A list with the updated workbook object (wb) and number of the last
row in the updated workbook object (length).
}
\description{
Adds references to an excel file
}
\examples{
wb <- openxlsx::createWorkbook()

openxlsx::addWorksheet(wb, "sheet")

fmt_ref(wb = wb, sheet = "sheet",
ref_text = ref_text,
start_row = 2, start_col = 1, end_col = 17)
}
