% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_title.R
\name{fmt_title}
\alias{fmt_title}
\title{Adds title to an excel file}
\usage{
fmt_title(wb, sheet, title_text, d, start_row = 1, start_col = 1, end_col)
}
\arguments{
\item{wb}{an openxlsx workbook object}

\item{sheet}{name of sheet in workbook object}

\item{title_text}{dataframe of text to display in title of workbook object}

\item{d}{dataframe of data to be included in table}

\item{start_row}{number for starting row}

\item{start_col}{number for starting column}

\item{end_col}{number for the last column}
}
\value{
A list with the updated workbook object (wb) and number of the last
row in the updated workbook object (length)
}
\description{
Adds title to an excel file
}
\examples{
wb <- openxlsx::createWorkbook()

openxlsx::addWorksheet(wb, "sheet")

fmt_title(wb = wb, sheet = "sheet",
title_text = title_text, d = sample_data,
start_row = 2, start_col = 1, end_col = 17)
}
